Whatsapp-Covid-Bot/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
argcomplete==1.12.0
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
boto3==1.14.22
botocore==1.17.22
branca==0.4.1
bs4==0.0.1
cachelib==0.1.1
cachetools==4.1.0
cfn-flip==1.2.3
click==6.7
colorful==0.5.4
constantly==15.1.0
cssselect==1.1.0
decorator==4.4.2
Deprecated==1.2.10
durationpy==0.5
earthengine-api==0.1.224
face-recognition-models==0.3.0
firebase==3.0.1
Flask==1.1.2
folium==0.11.0
future==0.18.2
futures==3.1.1
google-api-core==1.19.0
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
goslate==1.5.1
graphviz==0.14
gunicorn==20.0.4
hjson==3.0.1
howdoi==1.2.1
httplib2==0.18.1
httplib2shim==0.0.3
hyperlink==19.0.0
importlib-metadata==1.7.0
incremental==17.5.0
insult==0.4
ipython==7.15.0
ipython-genutils==0.2.0
itemadapter==0.1.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jwcrypto==0.7
kappa==0.6.0
lxml==4.5.1
markovmeme==0.0.1
newsapi-python==0.2.6
notify2==0.3.1
oauthlib==3.1.0
opencv-python==4.2.0.34
parsel==1.6.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip-tools==5.2.1
placebo==0.9.0
pprint==0.1
prettyprinter==0.18.0
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDictionary==1.5.2
PyDispatcher==2.0.5
PyGithub==1.51
PyHamcrest==2.0.2
pyjokes==0.6.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyquery==1.4.1
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.6.1
python-docx==0.8.10
python-firebase==1.2
python-http-client==3.2.7
python-jwt==3.2.6
python-slugify==4.0.1
python-telegram-bot==12.7
PyYAML==5.3.1
queuelib==1.5.0
regex==2020.6.8
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
Scrapy==2.2.1
secure-smtplib==0.1.1
sendgrid==6.4.1
service-identity==18.1.0
simple-imshow==1.0.post1
starkbank-ecdsa==1.0.0
text-unidecode==1.3
thug-memes==0.1.2
toml==0.10.1
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
troposphere==2.6.2
twilio==6.44.1
Twisted==20.3.0
twython==3.8.2
tzlocal==2.1
uritemplate==3.0.1
w3lib==1.22.0
wcwidth==0.2.4
webwhatsapi==2.0.5
Werkzeug==0.16.1
wget==3.2
wikipedia==1.4.0
wrapt==1.12.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
zope.interface==5.1.0
